feat: Improved threat and attack assessment
This commit extends qw's ability to assess the threat of monsters in
terms of buffs currently on the monster as well as on player attacks and
relevant buffs the player has, and monster resistances to player attack.
It also adds a system that compares attacks based on per-monster damage,
per-monster accuracy, attack delay, and any available player buffs we
both currently have or *could* use. Monster threat is calculated
relative to our primary attack and current buffs. Then, for monsters
that continue to have high threat, we use attack comparisons in an
ordered list to choose the most appropriate player attack. Currently the
only possible attacks to choose between are the primary and any throwing
attack. A subsequent commit will use the attack comparison system to
incorporate usage of most attack wands.

Going into a bit more detail, a monster's threat is increased by 50% if
it's hasted, by 50% if it's mighted (stacking with haste), or by 237.5%
if it's berserked (representing a factor of (3/2)^3). Finesse decreases
a monster's threat by half, haste by 2/3, and might and heroism decrease
it by an amount calculated from the ratio of damage (for might) or delay
(for heroism) compared to an attack without those buffs.

For high threat monsters, we only target these if either our primary
attack would already target this monster, or if the best attack could
hit both the scary monster and our primary target at the same time (e.g.
with a penetrating or explosive attack). This is to avoid letting
adjacent/reaching monsters get lots of hits as we try to deal with a
scary monster further away.

This is a large commit containing a lot of functionality because it's
difficult to split up into further commits. It's reasonably well tested
and bug-free at this point based on lots of local games, but will
continue to receive refinement and corrections.