feat: Use attack wands
This commit has qw use attack wands on scary monsters (threat 3 relative
to our primary attack with current buffs) when a wand has better damage
than our primary attack. The wands supported are, in order of
consideration:

flame, mindburst, iceblast, acid, light, quicksilver, and paralysis

The first wand that improves the monsters threat (by having higher
relative damage) to below 3 will be used, otherwise if no attack can do
so, the one with the best effective damage relative to the monster is
chosen. Both damage and accuracy are relative to the specific monster,
with the hex wands using the success chance versus that monster as
accuracy. For paralysis, we use the monster's remaining HP as the damage
done, with the reasoning that a paralyzed monster is essentially dead.
It might be good to give qw better ranged target prioritization for
paralyzed monsters in the future, but it already does this for melee
(via "stabbability").

Additionally we'd like to support other evocables like wands of roots
and charming, as well the lightning rod and possibly other summoning
evocables as well. For charming, we need a means of evaluating
effectiveness that looks at e.g. nearby monster targets and possibly the
threats of those targets. For roots, we'd need some kind of damage
calculation and probably to incorporate a consideration of the movement
prevention it applies. Summoning evocables would in general benefit from
a more precise way of evaluating the strength of potential summons.

We don't currently train evocations to support these wands better in
later parts of the game. We'd probably want to keep any evocations
training to be minimal until we had a better means of evaluating
relative effectiveness.

Some preliminary testing shows qw making good use of these wands with
early difficult monsters such as uniques.