feat: Simplify memoization
When memoizing a function that takes argument, instead of calling the
function with those arguments, assume the provided function is a closure
with any needed argument values bound so that it already calls the
function its memoizing with the right arguments. This means we don't
have to unpack arguments ourself, nor do we need to define those
arguments in the closure function. Requiring the closure introduces one
more guaranteed function call, but this trivial function call is faster
than the one that unpacks the arguments. And the closure definition is
less confusing to read as well.