feat: Improve feature state handling
Check feature state in c_persist during initialization, fixing up any
missing values. This is already done when updating state, but that's not
a good of a time to fix things up, so remove those cleanups. Doing
c_persist cleanups at initialization time is a soft form of "save
compat" for qw when we need to add new data and test on in-progress
games, something we're doing more frequently these days.