feat: Scripts to run qw in parallel
The run-qw.sh script can run a specific of iterations of qw, keeping
track of its progress in a temporary file. It makes backups of c_persist
and save data and exits when it sees that a crawl process hasn't removed
the .cs save file. This makes it easier to debug errors or use qw save
goals when testing new functionality. It makes symbolic links to a
provided rc file under a name provided as an argument, so we that each
invocation of this script can have its own c_persist data when it's run
in parallel.

The batch-qw.sh script runs a specified number of instances of run-qw.sh
in parallel, showing the output in tmux with a window for each instance.