feat: Support for Djinni
Update all aspects related to mp consumption and Gspirit to have logic
for Djinni. As Djinni, we check hp instead of mp for using mp, we don't
consider gods with MP-using invocations to actually do so (relevant to
antimagic, potions of magic, and Gspirit), we don't value Gspirit
positively. There are also a few weird cases like when to recall, which
likes to wait until we have max mp, and I've updated that to consider hp
instead, but that's only relevant to future usage of Yred and Beogh.