feat: Improvements to fleeing and retreating
This commit unifies the handling of threat for fleeing, teleporting, and
retreating. If threat is very high even with all available buffs for our
primary attack, we attempt to retreat to a position where no more than 2
monsters can attack us, or we flee to stairs, or finally, if neither of
those are possible, we teleport. This unifies the of approaches we have
to deal with very high threat levels into a well-defined system. A
subsequent commit will implement taking alternative stairs when threat
levels at arrival stairs are very high.

This commit also applies this logic to handling titanic and enormous
slime creatures, with the condition that we must either flee or teleport
from these. Slime creatures have their threat adjusted by the
scary_enemies table based on player xl and available buffs, so that they
can trigger the flee/teleport threshold only when the situation is
threatening enough. In the future, it would be good to handle all
greater-than-usual high threat single monsters this way, particularly
for very nasty monsters in the early game, but this would require the
goal system to have threat-based adjustment of both stair usage and
level exploration.