feat: Simplify qw batch scripts
Instead of creating and deleting symlinks in the provided crawl
directory, assume any necessary RC files will be found at the files in
the provided arguments for the RC file and the RC dir. This lets the
user set up and manage their own RCs in designated places without
messing up anything in their crawl directory.