feat: A plan to abort taking safe stairs
If we're unable to follow through with taking safe stairs, signal that
we need to abort, update our goals to reflect this, then restart the
plan cascade. This prevents us from being stuck when we try to take safe
stairs that we currently can't reach. Ideally we'd have more accurate
tracking of reachability between all stairs, but that requires more
work, including changes to the travel search algorithm.