fix: Accept a beam attack prompt for self-hits
When zapping things like a wand of flame in a direction that could hit
crystal walls, say yes to any resulting attack prompt, allowing the
attack to proceed.

Ideally we'd detect when this could happen ahead of time in ranged
attack code, but `spells.path()` doesn't continue the ray into any wall
it would hit. Rather than attempting some hacky prediction of the ray's
next position, we should update the clua API to at least have an option
for including the position of any wall that the ray might hit. I'm not
going to do this for now. This is only relevant to the rare situations
where a wand of flame is used near crystal walls, so the benefit is very
small.