fix: Disable retreating when using melee
I'd hoped to have retreating developed for melee for this release, but I
spent a lot of time on other tactics to improved ranged, and we're far
from complete there as well! I'd like melee to have proper consideration
of ducking out of los over ranged monsters, and we still have issues to
resolve with retreating and tactical steps.