feat: Coglin support, equipment set evaluation
This commit redesigns equipment evaluation to consider the cumulative
value of an entire set of equipment and evaluates all possible
combinations of equipment to find the set with the highest total value.
It uses this to support the new Coglin species in trunk.

To evaluate an item set, we tally up the independent aspects of each
item first, namely utility for base properties specific to the item
class (e.g. weapon damage) and any linear properties. Then, for
non-linear properties, we consider the final (capped) property level
obtained from all items in the set and get the utility for that level.

To find the best item set, we generate all combinations of items that
fill all available slots. For the set found with the highest value, we
remove the worst item in terms of the value it adds to the set if that
would increase the value of the set, repeating this process until the
remaining items all add value to the set. This allows to avoid the
complexity of "no item" as a slot choice and consider fewer combinations
overall.

The best equipment set is cached so that it's not recalculated each turn
until some relevant aspect changes, like picking up a new item, gaining
an experience level, skill level, a stat point. Item equipping and
dropping are centralized and simplified. Swapping terrible jewellery to
avoid 0-stat now only considers rings, since amulets are no longer
swappable.

For coglins, we use equipment set evaluation to accurately choose the
best pair of one-handed weapons. For simplicity, we never consider using
two-handed weapons on coglins. Since we swap weapons slowly, the hydra
scary monster logic is improved and can accurately choose to use
throwing attacks over melee attacks if our melee attacks can cut hydra
heads.

For scroll acquirement, Okawaru gifts, and coglin gizmos (chosen at XL
14), we use equipment sets to choose the best item. For each offered
acquirement item, we find the best equipment set using that item and our
current equipment. We then choose the acquirement item that has the
highest valued set among these best equipment sets.

This is a large commit that contains new functionality for attack
evaluation that would ideally be in separate commits, but it would be a
lot of work to split that into independent changes. Here's a list of
attack changes that had to be included in this commit.

* Throwing attacks are now properly evaluated for purposes of choosing
  the best player attack for scary monsters. This form of throwing
  attack always uses the missile with the best damage.
* Throwing is used as a secondary attack for monsters out of melee range
  only when we have at least a moderate level of threat. This form of
  throwing attack always uses the missile with the most ammo available.
* Unarmed attacks are properly evaluated when choosing a best attack for
  scary monsters.