fix: Travel first direction finalization
When finalizing the travel depth to check unexplored stairs on adjacent
levels, we weren't finalizing the `first_dir` value as well. This could
lead to qw getting stuck in a loop taking stairs when e.g. exclusions
forced it to use the first direction information. This commit finalizes
the first direction data at the same time depth is finalized. Also don't
set `first_dir` if we're already on the same level as our final depth,
so that this value is always about reaching the right depth and isn't
redundant with e.g. `stairs_dir`.