feat: Improvments to tactical step
Add a tactical step for putting out sticky fire, which has us
preferentially move to water or just to a nearby square if we're on
fire. Also improve the logic for slimy walls to consider the actual
slimy wall count at each potential movement destination, and to
recognize that we don't care about slimy walls for certain attacks.