feat: Improvements to fleeing and retreating
This commit improves retreating and fleeing conditions, fixing a number
of bugs. Monster and player move speed calculations are more accurate,
using the better information we have from clua now, and we use aut/delay
as a common scale for both monsters and player.

Retreating is brought back into tactical steps so that we can more
accurately control the interactions between retreating steps and other
types of steps. Additionally we consider player move speed for allowed
retreat distance range, so that e.g. spriggan is allowed to search for
positions further away.

Fleeing now uses the `best_move_towards()` system to get the most
accurate movement and tests for viability of movement. When fleeing from
extreme threats, we allow up to 2 monsters to make opportunity attacks
against us, instead of requiring that none can do so. Additionally, a
number of fleeing bugs have been fixed and we have better debugging in
the "flee" channel to help identify problems in the future.