feat: Increase qw memory requirements
I've been running into memory issues in a small number of games in Lair
and Vaults, where levels are larger. So this commit increases the
default clua memory limits from 96MB to 128MB. Once again, I should add
memory debugging functionality and investigate this to see if we have
memory leaks, but increasing this limit works for now.

Also remove the MAX_MEMORY_PERCENT rc variable and just use the
MAX_MEMORY variable as the limit. This variable can be set to a value
less than the available memory to prevent out-of-memory crashes.
Document this variable with some further details in the rc file.