feat: Better movement in clouds and towards runes
This commit reworks cloud movement so that we wait for damaging clouds
to dissipate if we're trying to move to a location with a dangerous
cloud that's not excluded. This prevents qw from killing itself when
reading a scroll of poison and allows use general plans to move to and
pick up a rune even when it's under clouds. The exclusion condition is
necessary to prevent waiting for cloud generators to dissipate. A future
update to this system could employ waiting even for excluded clouds
using a capped wait timer.

The safety logic has been improved and generalized so that we can
accurately determine whether specific clouds and traps are safe without
having to detangle that information from `view.is_safe_square()`. Also
fix a debug `false` added to a condition that prevented movement towards
safety from working.