feat: Tweak a melee waiting condition
We wait for monsters to reach us in melee when possible, but don't do
this if a monsters has any sort of ranged attack, including reach
attacks. It would be ideal to wait for monsters whose only ranged attack
was reach until that monster was in reach range. I'm tweaking the
condition in `Monster:player_can_wait_for_melee()` to allow this, but
until we have a monster flag or other clua means to accurately check for
ranged that properly excludes reach, it's not wise to use this
functionality. Hence, for now, I'm keeping the condition in
`plan_melee_wait_for_enemy()` that prevents waiting for any sort of
ranged monster.