feat: An implementation of kiting
Spriggans need kiting techniques to take full advantage of their fast
move speed. This commit implements a kiting system for qw that considers
the various factors of player/monster move delay, player attack delay,
monster locations, and nearby terrain.

In order to engage in kiting instead of either retreating or fighting
from the current position, some conditions are required:

1. Our HP as at least 50%, we are not confused, nor in the Abyss.
2. We have a ranged attack target. This may be for our primary launcher,
   a throwing attack, or a wand.
3. The total threat is at least 5. This is the same cutoff used for when
   to consider retreating.
4. There are no ranged attacking (ignoring reach) nor "los danger"
   monsters such as dream sheep, glass eyes, torpor snails etc that we
   specifically target regardless of proximity whenever these monsters are
   in LOS.
5. For any monster that is close enough where a kiting step is warranted
   (see below), that monster (a) is slower than us and (b) it would not
   take 6 steps or fewer to gain a distance of one on the monster.

A kiting step is considered necessary if the use of our intended attack
gives the monster enough time to get into melee range.

To determine a kiting step, we filter and score nearby destination
positions in a LOS distance radius. Destinations that are less than 3
squares away, or that aren't reachable, or that would reveal new LOS
tiles are excluded. The remaining destinations are scored by the three
criteria described below, taking the sum of scores for all monsters in
LOS, with each monster's score multiplied by its threat. The criteria
are:

1. avoidance: The distance between the first move we would make
   towards the destination and the last position a monster would arrive
   at in melee range of our current position. The large the distance
   between these two positions, the more the destination requires is to
   move "away" from a monster we're trying to kite. For reach monsters that
   can already melee us, we add half a point if they can't reach attack
   us at the first move position.
2. visibility: The number of monsters that can see the first move we
   would make towards the destination. More monsters being able to see this
   first move means fewer cases of monsters taking shortcuts that put them
   in melee range.
3. distance: The number of moves it takes the monster to reach the
   destination minus the number of moves the monster could make in the
   time it takes us to reach that destination.

All destinations are compared first on avoidance, then by visibility for
ties in avoidance, then by distance for ties in visibility, in order to
get a best destination. For the best destination, We take the
aforementioned first move towards that destination as the kiting step.
If no kiting step is needed, we attack the current ranged target until
another kiting step is needed or combat ends.

In practice, this algorithm seems to do an decent job in balancing the
various aspects needed to kite safely while keeping monster movement on
track so the kiting process can proceed. The kiting algorithm could see
refinement in a number of ways, such as allowing kiting of ranged
monster based on their attack range, coordination of kiting with stairs,
and considering "loops" in terrain to get better kiting destinations.