refactor: Standardize some variable names
To help avoid errors where a key of `distance` is expected by `dist` is
actually used (or vice versa), I'm standardizing use of "distance" for
any global function or data and "dist" for anything that's local or just
a return value. Here global means either actually global or directly in
the qw or const tables. A modern IDE can probably alleviate most of the
need for attempting variable name standarization, but the realities of
working both in lua and in an a script executed in the clua environment
mean that some standardization can reduce the chance for certain bugs.