feat: No outnumbered tactical steps for ranged
The "outnumbered" tactical is a simple way to try to improve positioning
versus adjacent and ranged monsters when not using cleave. For ranged
weapons, we have the more recent retreating system, which is able to
analyze terrain and monster movement more thoroughly. So disable
outnumbered steps in that case as well. In the future, we'll extend
retreating to melee (both with and without cleave), so it's likely the
outnumbered step will go away entirely.