feat: threat level adjustment by XL
At early XLs, qw needs to be more willing to kite, retreat, and use
buffs. To some extent this can be handled via having monsters designated
as nasty based on XL. Ideally we could better assess monster threat
based on our defenses and primary office, but that requires more
significant changes. For now, this commit tries a simple approach of
decreasing threat thresholds for lower XL, increasing it to the
previously used thresholds by XL 10.

We make the thresholds as follows:

* Moderate: 2 (XL1-7) / Increase by 1 for each XL from 8 to 10
* High:     5 (XL1-7) / Increase by 2 for each XL from 8 to 10
* Extreme:  10 (XL1-5) / Increase by 2 for each XL from 6 to 10

Hopefully this gets qw to consider retreating, buffing, and fleeing a
bit more in the early game, when it's especially vulnerable.