feat: Distance map move searches
Previously we could only easily apply the checks in a square function to
the next immediate move given by a distance map, but couldn't check the
entire path. Hence for something like fleeing, we had limited
information about whether our distance mapped movement is likely to
succeed given our current position. This commit implements move searches
that use apply a square function to a distance map to check whether a
move path can be found that satisfies the square function at every
position. This is mostly the same process used by the `move_search()`
algorithm, except we're using the distance map to check for next
positions that decrease our distance to the target. The benefits of this
approach will be immediate for fleeing, but I'll probably replace some
code used for retreating with this algorithm as well.

To that end, distance map move searches allow passing in an external
cache that can be used across searches with different centers but the
same target. When this isn't passed in, one is initialized and used
separately for each individual search in the same way the current
`move_search()` attempt cache currently works. However it also tracks
successes, recording the `last_pos` result, in addition to failures,
which are cached as `false`. So in the future, we can improve the
efficiency of retreating searches by sharing a cache among all searches
for retreat positions that search to the player's position as a target.