feat: Throttle best equip set choice
As runs of OpHu have shown, qw can sometimes carry over a dozen rings,
which makes the number of equipment combinations to consider go well
into the thousands. It requires a lot of calculation to run through
these, which will have to be addressed in a number of ways. This commit
helps prevent the WebTiles server from seeing qw's crawl process as
being stuck in an infinite loop but using the coroutine throttle and
using a delay after each block of 100 equipment sets. This works well in
practice to prevent qw's crawl process from getting killed. However with
14 rings, 2 weapons, a shield, and a hat it, takes qw approximately 3
minutes to run through all the equipment sets running on CDI.

It may be somewhat rare for a qw Op game to have that many rings at once
in a complete game, but it's certainly not the worst case. To avoid qw
taking hours just to compute a best equipment, we'll need to work on (1)
dropping more items that we can see will never be realistically useful
(2) reduce the amount of computation when calculating equipment sets
through cacheing, and (3) look into ways to limit the number of
equipment sets we calculate. For now, hopefully this commit is enough to
at least let qw complete its calculations.

We also move a top env variable into the qw table and remove a throttle
variable that's not actually needed.