feat: Don't reset equipment set cache for skills
We currently track skill levels and reset the best equipment set
whenever a skill level has increased by 1. I added this when
implementing equipment set caching in case we move to considering skill
levels for equipment decisions in the near future. We're not doing that
in the near term, and we're currently sometimes struggling with
computing equipment sets for octopodes with a lot of rings. So for now,
let's disable resetting best equipment sets due to skills but keep the
skill tracking code in place.