fix: Improve a check for player melee
For `Monster:player_can_melee()`, add an `ignore_temp` argument to
ignory temporary conditions that prevent melee, such as being caught in
a web or net. Use this argument for the call to this method in
`Monster:player_has_path_to_melee()`, which is used when determining
whether to exclude a monster. Otherwise we end up excluding monsters
already in melee range just because we temporarily can't melee them.