Fix SK_DODGING/SK_SHIELDS formulas in skill valuation code
The dodging formula was outdated and the shields formula didn't make sense
(the value should not depend on current shields skill). The new shields
formula tries to account for the three things that SK_SHIELDS does (SH, EV,
attack delay reduction) in a somewhat principled way. SH gain is weighted
by 0.8 instead of 1.0 because qw uses that ratio for SH vs AC/EV elsewhere.
Delay reduction is weighted by 2.4 because we currently value raising
your weapon skill by 2 at 3.0, and a large part of that is reducing delay
reduction by 1.