Decrease value of some attributes when you have a lot of them already
In Crawl, most things have diminishing returns at some point - for example
if you have lots of EV then you will already be relatively good against
attacks that check EV but relatively bad against attacks that ignore EV,
so getting more EV is less effective at decreasing damage taken. With this
in mind, we start decreasing the value of each of AC, EV, and SH once you
have 25 of them. We also start to decrease the value of raising Unarmed
Combat skill once you have 18 of it - this is partly because extremely
high values of UC skill are less important because you are already doing
tons of damage, and partly because the formula is a bit different than
weapons.