feat: Improve qw messaging
Change the message prefix from "|||" to "QW:", since the latter is
clearer and doesn't cause formatting issues when pasted into various
places like Discord that "|" in special ways. Centralize this prefix so
that changing it only involves updating one place in the code.

Also, move the turn count number into this prefix, so we have a prefix
like "QW(NNNN):", instead of printing it before the prefix. Finally,
don't include the turn count when making notes, since that's already
included in the note table formatting.