fix: Minimum distance for movement searches
We were incorrectly considering only grid distance when determining
whether a movement search should end based on reaching the minimum
required distance to its target. This prematurely terminated certain
movement searches with a minimum distance of two when the destination
was two squares away, but on the other side of a wall.

This commit requires that the final search position can melee the target
position when the minimum distance is greater than 0. Since movement
searches currently only happen either in the context of directly
reaching the target square (min distance 0) or reaching a melee position
for the target square, this fix works for now. In the future, if we need
movement search for some other reason, we probably want to support
passing an arbitrary destination square function that could check
whatever arbitrary conditions we needed (e.g. line of fire).