feat: A debug log that can be written to a file
This commit adds a debug log that can be enabled with enable_debug_log()
such that most dsay() messages are appended to the table qw.debug_log
instead of being printed to the message log. In cases where large
amounts of debug messages are generated, it can be impractical to print
them all. The log can be inspected with print_debug_log(), which prints
a slice of the log, or written to a file with write_debug_log(). The
latter function requires that crawl be built with
-DCLUA_UNRESTRICTED_LIBS so that we have access to the io library.