fix: Properly cache nil results
For Monster:melee_move_search(), we weren't caching a nil result as
false, which meant that it wasn't being cached in that case. Also
simplify the caching implementation of this and other monster methods
that use a position argument by using the position hash as a key instead
of the individual coordinates in two arrays.