fix: Unsafe distance mapped moves
We weren't checking carefully for the case when our best distance mapped
move was unsafe. When we subsequently found a move that's safe and
should therefore have priority over the current best (unsafe) move, if a
safe move had worse distance, it wouldn't be chosen, leading to
potential infinite walking loops. Now we always prefer a safe move
regardless of distance when our current best move is unsafe.