fix: Logic for enchantment scrolls
Now that best equipment choices involve the throttle, we can't have
uncached equipment set choosing in our c_choose... functions for good
consumables, since those aren't run via a coroutine. We can always use
the cached result when reading a known scroll, but weren't taking
advantage of this due to how the logic indicating whether a scrolls was
known worked. This commit uses variables in the qw table to indicate
that weapon that we'd like to target for known enchante weapon/armour
and brand weapon scrolls.  This allows us to avoid performing new (and
hence uncached) equipment set searches for e.g. fallback weapons when
we'd never need one.

There is still a potentially problematic case when we read unknown brand
weapon and our weapon in the best equipment set can't or shouldn't be
branded. This hasn't come up in practice since we'd rarely have enough
equipment items to trigger the throttle in this case, but it should be
addressed in the future. The best solution is likely to wrap the current
logic of the c_choose_... functions into coroutines that these functions
then call, which would allow us to use the throttle.