fix: The condition for using secondary throwing
The condition for when to use our throwing attack with the most
ammuniation was inverted. Also, lower the threat threshold for using
this secondary throwing attack from 3 to 2, so that we're always willing
to use it for threat level 2 monsters. Although we should feel safe
taking on such monsters 1v1, there's little reason not to soften them up
before melee starts.