feat: Better fleeing on the orb run
With the new melee retreating code, it's more likely for melee qw to
move a away from its position at the start of combat in order to break
LOS, which can lead to more protracted combat and more spawns. Normally
it's better to fight fewer monsters at once and without LOS of ranged
attackers whenever possible, but in the case of the orb run, we really
want to press on to D:1 unless combat is inevitable.

This commit has qw always try to flee from combat when it has the orb if
it satisfies our existing conditions about not having close following
attackers. These are based on distance calculations that avoid
opportunity attacks as well as many ranged attacks from monsters that
are relatively close, and these haven't change in this commit. I've also
been conservative by not changing retreating itself, since qw will
inevitably have to fight on the orb run, and we don't want it to do so
from bad positions. That said, probably some fleeing distance and
retreating calculations should change for the orb run. I've tested the
orb run several times on a qw character that recently died with the orb
in a live game, and the results look improved, so this is all I'm doing
for now.