refactor: Use position hashes for map indexing
The traversal, exclusion, and distance mapping systems currently use a
doubly-indexed array using a position's coordinates. This is cumbersome
both because the first array has to be initialized and because any
indexing has to check that the coordinates are within bounds of the
initialization. This commit refactors the code to have all map arrays
use a position's hash as its coordinates. This requires the use of only
one table and means we never have to check coordinate bounds.

Additionally we remove the adjacent floor map cache, since the retreat
system will no longer use this.