feat: Improved throwing and ranged assessment
This commit fixes a few bugs in ranged attack assessment and refactors
the ranged code. Some fixes are related to more subtle aspects of
explosion assessment or more accurate assessment of monster alignment.
One important fix is for determining if an attack could hit a monster
for purposes of selecting the strongest attack. With this commit, we use
the actual functions that try to select a target for the attack in
question to see if one can actually be found. Previously we had
independent line of fire checks that aimed only directly at the target
and didn't consider all possible firing locations.

Additionally, this commit improves the use of throwing. We now add a
penetrating throwing attack (i.e. javelins) to the attack list if our
highest damage throwing attack is non-penetrating, so that we can
potentially hit high threat targets behind other monsters. The moderate
threat condition for using the secondary (highest ammunition count)
throwing attack has been fixed; its logic was inverted. We also use
boomerangs when available even if we don't have moderate threat levels.
Boomerangs last a long time and are appropriate to spam against monsters
with threat less than 2 while waiting for monsters to enter melee range.