feat: Improve assessment of Okawaru's Heroism
This commit gives qw the ability to see when Heroism would have minimal
impact. We now look at the result that Heroism would have on our best
attack on a scary enemy, or on our primary attack if not dealing with a
scary enemy, in terms of delay and damage (for UC and Throwing). We
don't use Heroism unless our overall damage for the attack in question
would increase by at least 20%.