Tweak some monster move search code
Rename Monster:has_path_to_player() to Monster:can_traverse_to_player()
to emphasize that this function only cares about traversal. Use this
function instead of the melee path search when assessing enemies so we
don't ignore the threat of monsters that are blocked by other monsters.
This assessment is trying to consider the entire duration of combat, and
all monsters will eventually have an open path.

Also relax the minimum distance required for the
:can_traverse_to_player() search from 0 to 1, so that we don't have to
perform an entirely new search for the incoming monsters check. A
monster reaching us at distance 1 should be able to follow us as we
retreat one step. If unexpected corner cases arise, we can either adjust
this limit or trying another approach that hopefully doesn't involve a
new move search.